requirements.txt
aioredis==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
Automat==22.10.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
channels==2.3.1
channels-redis==2.4.2
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==40.0.2
daphne==2.5.0
dj-database-url==2.0.0
Django==3.2.5
django-appconf==1.0.5
django-bootstrap-form==3.4
django-bootstrap4==23.1
django-user-accounts==3.2.0
djangorestframework==3.14.0
gunicorn==20.1.0
hiredis==2.2.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==0.6.2
pinax-templates==3.0.0
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.1.1
pytz==2023.3
requests==2.30.0
service-identity==21.1.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
urllib3==2.0.2
whitenoise==6.4.0
zope.interface==6.0
